Move ClipUnits into GL.Shader.DSL.
This is probably not a great ultimate home for it but it’ll have to do for now.
